+++ eufs_rqt/src/eufs_rqt/EUFSRobotSteeringGUI.py
@@ -350,20 +350,20 @@ class EUFSRobotSteeringGUI(Plugin):
             - self._widget.angular_slider.singleStep())
 
     def _on_max_linear_changed(self, value):
-        self._widget.linear_slider.setMaximum(
-            value * EUFSRobotSteeringGUI.slider_factor)
+        self._widget.linear_slider.setMaximum(int(
+            value * EUFSRobotSteeringGUI.slider_factor))
 
     def _on_min_linear_changed(self, value):
-        self._widget.linear_slider.setMinimum(
-            value * EUFSRobotSteeringGUI.slider_factor)
+        self._widget.linear_slider.setMinimum(int(
+            value * EUFSRobotSteeringGUI.slider_factor))
 
     def _on_max_angular_changed(self, value):
-        self._widget.angular_slider.setMaximum(
-            value * EUFSRobotSteeringGUI.slider_factor)
+        self._widget.angular_slider.setMaximum(int(
+            value * EUFSRobotSteeringGUI.slider_factor))
 
     def _on_min_angular_changed(self, value):
-        self._widget.angular_slider.setMinimum(
-            value * EUFSRobotSteeringGUI.slider_factor)
+        self._widget.angular_slider.setMinimum(int(
+            value * EUFSRobotSteeringGUI.slider_factor))
 
     def _on_strong_increase_linear_pressed(self):
         self._widget.linear_slider.setValue(
